pkg/updateengine: move integration tests to separate file
So they run behind a build tag, rather than with the t.Skip() gate.
